+++ R/getCIMLR.R
@@ -25,6 +25,15 @@ getCIMLR <- function(data        = NULL,
                      cores.ratio = 0,
                      verbose     = TRUE){
 
+  # check data
+  n_dat <- length(data)
+  if(n_dat > 6){
+    stop('current verision of MOVICS can support up to 6 datasets.')
+  }
+  if(n_dat < 2){
+    stop('current verision of MOVICS needs at least 2 omics data.')
+  }
+
   useless.argument <- type
   if(verbose) {
     fit <- quiet(CIMLR(data,