+++ R/getConsensusClustering.R
@@ -48,6 +48,15 @@ getConsensusClustering <- function(data         = NULL,
                                    seed         = 123456,
                                    verbose      = F){
 
+  # check data
+  n_dat <- length(data)
+  if(n_dat > 6){
+    stop('current verision of MOVICS can support up to 6 datasets.')
+  }
+  if(n_dat < 2){
+    stop('current verision of MOVICS needs at least 2 omics data.')
+  }
+
   useless.argument <- type
   if(is.null(norMethod)) {
     d <- do.call(rbind, data)