+++ R/getLRAcluster.R
@@ -23,22 +23,33 @@ getLRAcluster <- function(data       = NULL,
                           type       = rep("gaussian", length(data)),
                           clusterAlg = "ward.D"){
 
+  # check data
+  n_dat <- length(data)
+  if(n_dat > 6){
+    stop('current verision of MOVICS can support up to 6 datasets.')
+  }
+  if(n_dat < 2){
+    stop('current verision of MOVICS needs at least 2 omics data.')
+  }
+
   data <- lapply(data, as.matrix)
 
   if(is.element("binomial",type)) {
     bindex <- which(type == "binomial")
-    a <- which(rowSums(data[[bindex]]) == 0)
-    b <- which(rowSums(data[[bindex]]) == ncol(data[[bindex]]))
-    if(length(a) > 0) {
-      data[[bindex]] <- data[[bindex]][which(rowSums(data[[bindex]]) != 0),] # remove all zero
-    }
+    for (i in bindex) {
+      a <- which(rowSums(data[[i]]) == 0)
+      b <- which(rowSums(data[[i]]) == ncol(data[[i]]))
+      if(length(a) > 0) {
+        data[[i]] <- data[[i]][which(rowSums(data[[i]]) != 0),] # remove all zero
+      }
 
-    if(length(b) > 0) {
-      data[[bindex]] <- data[[bindex]][which(rowSums(data[[bindex]]) != ncol(data[[bindex]])),] # remove all one
-    }
+      if(length(b) > 0) {
+        data[[i]] <- data[[i]][which(rowSums(data[[i]]) != ncol(data[[i]])),] # remove all one
+      }
 
-    if(length(a) + length(b) > 0) {
-      message(paste0("remove a total of ",length(a) + length(b), " features because their categories are not equal to 2!"))
+      if(length(a) + length(b) > 0) {
+        message(paste0("--", names(data)[i],": a total of ",length(a) + length(b), " features were removed due to the categories were not equal to 2!"))
+      }
     }
     type[bindex] <- "binary"
   }