+++ NEWS.md
@@ -3,3 +3,8 @@
 * Minimal version of R was declined to 4.0.1.
 * Two functions were added. 
 * Troubleshooting section was updated.
+
+# MOVICS 0.99.6
+
+* Add an error notification prompt: current version of MOVICS can support up to 6 omics datasets and at least 2 datasets must be provided.
+* Fix bugs when providing more than one binary matrix.